requirements.txt
Pillow>=6.0.0
pygame>=1.9.6
